重新组织文件结构
